Remove legacy Propagator and its supporting code
The dispatcher pipeline is now the production path. Everything in the
legacy propagation service is dead code:

  - propagator.go: the parallel-consumer Propagator type and its
    flushBatch / processBatch / registerBatch / broadcast helpers.
    Replaced by Dispatcher + dispatcherConsumer +
    dispatcherBroadcaster.

  - reaper.go: the lease-coordinated PENDING_RETRY reaper.
    Replaced by the dispatcher's in-memory retry queue (Kafka replay
    provides crash durability).

  - retryable.go: IsRetryableError text-match classification.
    Replaced by isRetryableStatusCode using the HTTP status codes
    Teranode now returns (per the audit PR landed upstream).

  - backoff.go: ComputeBackoff helper for the reaper. The dispatcher
    has its own inlined retry-deadline math (computeRetryDeadline).

  - replay.go: the one-shot merkle-service replay on startup. Useful
    as a recovery mechanism for merkle-service state loss but not
    needed for the initial cutover; a follow-up can add it back if
    operationally warranted.

  - propagator_test.go: 2k lines of legacy-path tests, including the
    comprehensive mockStore with bookkeeping for PendingRetry,
    retryCount, merkleMarks, etc.

  - health_test.go: tests of the legacy propagator's broadcast
    behavior (first-success-wins across endpoints). Equivalent
    coverage now lives in dispatcher_broadcaster_test.go for the new
    broadcaster.

The propagationMsg type moved from propagator.go to dispatcher.go
where it's the natural neighbor of the dispatcher that consumes it.

A minimal mockStore lives in mocks_test.go — just enough to satisfy
the store.Store interface via embedding. Tests that need specific
store behavior compose over it (broadcastTestStore, pipelineTestStore).

Net: 4276 lines removed, ~25 lines added.

Verified:
  - go test -count=1 -race ./... -- all packages pass
  - golangci-lint run ./services/propagation/ -- 0 issues
